BigNumber with Time Comparison:
- Extract comparison period calculation to an util file and make use of it when building the query. This is to support following work that will make use of that function and to prepare us when adding proper testing.
